fix up unit tests since common functionality was removed from PortManager into utilities
